+++ zorphy_migrator/test/fixtures/exchangeable_enum/expected.dart
@@ -1,0 +1,11 @@
+enum NavigationType {
+  ///A link with an href attribute was activated by the user.
+  LINK_ACTIVATED,
+  ///A form was submitted.
+  FORM_SUBMITTED,
+}
+
+enum JsAlertResponseAction {
+  ///Confirm that the user hit confirm button.
+  CONFIRM,
+}